requirements.txt
streamlit==1.42.0
PyMuPDF==1.25.3
python-docx==1.1.2
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy
